refactor(cli): Simplify per review — drop unneeded guards, scope DIRECT_DATABASE_URL to Neon only
Per review feedback, several pieces of defensive machinery were removed as
not worth their complexity for this command:

- Dropped the missingPrismaConfig / unrecognizedPrismaConfig checks and
  their tests — prisma.config is generated by the framework and isn't
  expected to be missing or malformed; if it somehow is, the resulting
  error is acceptable without a dedicated guard.
- Dropped the hasSqliteUsageOutsideDb scan (api/src and scripts/) — nice to
  have, but not required, and it was a meaningful chunk of the diff for
  code and tests that exist to protect against an edge case.
- Reverted the datasource-URL rewrite from a comment-aware, line-based
  parser back to a single anchored regex replace — simpler, and the
  comment-defeating scenario it was hardened against isn't a case this
  command needs to fully defend against.

Also re-scoped DIRECT_DATABASE_URL: setup commands run locally with all
connection info expected in .env. A generic Postgres provider only hands
out one connection string, so `setup database postgres` no longer touches
prisma.config at all — it stays on DATABASE_URL. Neon is the one provider
here that hands out a separate pooled/direct pair, so the prisma.config
rewrite (DATABASE_URL -> DIRECT_DATABASE_URL) moved out of the shared
getSqliteToPostgresTasks and into neonHandler.ts, where it's actually
needed.
